+++ aim/web/ui/src/components/ImagesPanel/index.ts
@@ -1,0 +1,4 @@
+import ImagesPanel from './ImagesPanel';
+
+export * from './ImagesPanel.d';
+export default ImagesPanel;