+++ aim/web/ui/src/pages/Metrics/components/SelectForm/SelectForm.tsx
@@ -52,7 +52,7 @@ function SelectForm({
   const searchMetricsRef = React.useRef<any>(null);
 
   React.useEffect(() => {
-    const paramsMetricsRequestRef = projectsModel.getParamsAndMetrics();
+    const paramsMetricsRequestRef = projectsModel.getProjectParams(['metric']);
     paramsMetricsRequestRef.call();
     return () => {
       paramsMetricsRequestRef?.abort();