+++ aim/web/ui/src/types/components/ChartPanel/ChartPanel.d.ts
@@ -38,7 +38,7 @@ export interface IChartPanelProps {
     activePoint: IActivePoint,
     focusedStateActive?: boolean,
   ) => void;
-  resizeMode: ResizeModeEnum;
+  resizeMode?: ResizeModeEnum;
 }
 
 export interface IChartPanelRef {
@@ -57,5 +57,6 @@ export type IMemoizedForwardRefComponent<T> = React.MemoExoticComponent<
 export interface IChartTypeConfig {
   [key: string]:
     | IMemoizedForwardRefComponent<ILineChartProps>
-    | IMemoizedForwardRefComponent<IHighPlotProps>;
+    | IMemoizedForwardRefComponent<IHighPlotProps>
+    | IMemoizedForwardRefComponent<any>;
 }